requirements.txt
python-telegram-bot[webhooks]==22.3
aiohttp>=3.9
openai==1.*
pandas
python-dotenv